Add print() controls to update_resource_info()
This method belongs to the zenodotus.py module.

The print() controls added here aim to follow the program closely, to
see if the right conditional statement is being processed or not when
receiving updated info on a deposit/record from the Zenodo API.

In case of a positive answer, Lycophron needs to update the database
with this very same information, and it's suppose to do so few lines
below.