Modify print() controls in update_resource_info()
This method belongs to zenodotus.py.

It modified the print() control statements. Instead of showing info be-
fore the  conditional statement of the 'try/except' statements, it shows
after the 'try', and therefore, right before actually updating the info
on the original database. It shows in this following format:

"Field: <field_name>, New data: <new_data>, Old: <old_data>"

The aforementioned conditional statement tests if the <new_data> is
different than the <old_data>. If so, it goes to the 'try' statement.